Quick & Dirty hack to add timestamps support
I know this can be done better, but I was just playing around with Mapper on a small project.

Usage:

If the model contains these fields:

* created_at
* updated_at

and the model was called like this:

   var Post = Mapper.map("posts", { timestamps: true } );

then :created_at and :updated_at are automagically timestamped.

That's it. Nothing fancy, but 20 minutes of hacking fun ;-)